HU03 home inventory

+++ app/src/main/java/com/univalle/widgetinventory/model/Producto.kt
@@ -3,6 +3,6 @@ package com.univalle.widgetinventory.model
 data class Producto(
     val codigo: Int,
     val nombre: String,
-    val precio: Int,
+    val precio: Float,
     val cantidad: Int
 )
